Start ACPI support implementation

+++ kernel/include/utils.hpp
@@ -26,7 +26,7 @@ const char* str_until(char* a, char c);
 const char* str_from(char* a, char c);
 
 template<typename T>
-void memcopy(T* destination, const T* source, uint64_t size){
+void memcopy(T* destination, const T* source, size_t size){
     --source;
     --destination;
